refactor: some changes before diesel_async
